[util] New structure `LazyMatrixProduct`
The new structure represents a matrix product of the form A_N A_{N-1} ... A_1.

Functions `pushfirst!()` and `popfirst!()` can be used to add and remove
multipliers to/from the left of the product. The product is lazy in the sense
that the actual multiplication takes place only when the `eval!()` function
is called. The structure keeps track of previously evaluated partial products
and reuses them upon successive calls to `eval!()`.

The matrix multiplication is carried out by `Octavian.matmul!()`.